+++ sketch-extensions/src/main/java/com/github/panpf/sketch/internal/ImageXmlAttributes.kt
@@ -214,6 +214,7 @@ private fun parseResizePrecision(value: Int): Precision =
         1 -> Precision.EXACTLY
         2 -> Precision.SAME_ASPECT_RATIO
         3 -> Precision.LESS_PIXELS
+        4 -> Precision.SMALLER_SIZE
         else -> throw IllegalArgumentException("Value not supported by the 'sketch_resizePrecision' attribute: $value")
     }